# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sqordia API",
    "description": "Comprehensive API for business plan generation and management platform with Google OAuth authentication support",
    "version": "1.0.0",
    "contact": {
      "name": "Sqordia API Support",
      "email": "support@sqordia.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.sqordia.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user information",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User information retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate with Google OAuth",
        "description": "Authenticate user using Google OAuth tokens. Creates new user if not exists, or links to existing user if email matches.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthRequest"
              },
              "example": {
                "idToken": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE2NzAyNzQ0NzQ...",
                "accessToken": "ya29.a0AfH6SMC..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Google token or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/link": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Link Google account to existing user",
        "description": "Link a Google account to an existing local user account. Requires authentication.",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkGoogleAccountRequest"
              },
              "example": {
                "idToken": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE2NzAyNzQ0NzQ...",
                "accessToken": "ya29.a0AfH6SMC..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google account linked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Google token or account already linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Google account already linked to another user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/unlink": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Unlink Google account from user",
        "description": "Remove Google account association from user account. User must have a local password to unlink Google account.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Google account unlinked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "User is not a Google user or cannot unlink",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "organizationName": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "value": {
            "$ref": "#/components/schemas/UserDto"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "value": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "user": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "example": false
          },
          "value": {
            "type": "object",
            "nullable": true
          },
          "errorMessage": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "GoogleAuthRequest": {
        "type": "object",
        "required": ["idToken", "accessToken"],
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google ID token from OAuth flow",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE2NzAyNzQ0NzQ..."
          },
          "accessToken": {
            "type": "string",
            "description": "Google access token from OAuth flow",
            "example": "ya29.a0AfH6SMC..."
          }
        }
      },
      "LinkGoogleAccountRequest": {
        "type": "object",
        "required": ["idToken", "accessToken"],
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google ID token from OAuth flow",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE2NzAyNzQ0NzQ..."
          },
          "accessToken": {
            "type": "string",
            "description": "Google access token from OAuth flow",
            "example": "ya29.a0AfH6SMC..."
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      }
    }
  }
}